feat: Eliminando interfaces

+++ src/position/position.repository.js
@@ -1,5 +1,4 @@
 const positionEntity = require('./position.entity')
-const positionInterface = require('./position.interface')
 
 class PositionRepository {
     async getAllPosition() {
@@ -12,15 +11,6 @@ class PositionRepository {
     }
 
     async createPosition(positionData) {
-        const isValidData = Object.entries(positionInterface).every(([key, expectedType]) => {
-            const actualType = typeof positionData[key]
-            return actualType === expectedType.name.toLowerCase()
-        });
-
-        if (isValidData.includes(false)) {
-            throw new Error('Dados inválidos para criação do usuário');
-        }
-
         try {
             const { name, base_salary, subsidy, net_salary } = positionData
 
@@ -44,15 +34,6 @@ class PositionRepository {
     }
 
     async updatePosition(positionData) {
-        const isValidData = Object.entries(departamentInterface).every(([key, expectedType]) => {
-            const actualType = typeof departamentData[key]
-            return actualType === expectedType.name.toLowerCase()
-        });
-
-        if (isValidData.includes(false)) {
-            throw new Error('Dados inválidos para criação do usuário');
-        }
-
         try {
             const { id, name, base_salary, subsidy, net_salary } = positionData
 
@@ -80,8 +61,10 @@ class PositionRepository {
     async deleteDepartament(positionData) {
         try {
             const { id } = positionData
-            const deleteDepartament = await positionEntity.destroy({ where: id })
-            if (deleteDepartament) return { message: 'Departamento eliminado com sucesso' }
+            if (id !== undefined && !isNaN(id) && id.toString().trim() !== '') {
+                const deleteDepartament = await positionEntity.destroy({ where: id })
+                if (deleteDepartament) return { message: 'Departamento eliminado com sucesso' }
+            } return { message: 'O campo é obrigatório' }
         } catch (error) {
             throw error
         }
